fix multi-az db constant replacement bug (#2845)
The API requires to specify multiple vswitches for provisioning
a multi-AZ database, the problem is that the Alibaba Cloud API
will only returns the first one when subsequently queried about
the state of the database.
This makes terraform believes there is always a change of configuration,
hence it always tries to replace the database.

With this fix, we just compare the first vswitch to know if there
was a change or not.