create new vswitch_id_slave_1 rds for multi-az db
It is currently possible to specify multi vswitches
in the `vswitchId` argument, unfortunately the additional
vswitches information is lost after initial creation as
RDS API doesn't return the information.
Hence it's impossible to detect subsequent changes later.

This new argument `vswitches_id_slave_1` is intended to
solve this by storing this information in a separate field,
which will allow terraform to detect changes based on
comparison with the previous state recorded by terraform.